requirements.txt
aiogram==2.25.1
flask
flask-cors
mysql-connector-python

